backend/requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
bcrypt==4.1.2
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
distro==1.9.0
Django==4.2.10
django-celery-results==2.5.1
djangorestframework==3.14.0
dnspython==2.6.1
docker==7.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
eventlet==0.35.2
exceptiongroup==1.2.0
greenlet==3.0.3
gunicorn==21.2.0
humanize==4.9.0
idna==3.6
iniconfig==2.0.0
jsonschema==3.2.0
kombu==5.3.5
packaging==23.2
paramiko==3.4.0
pluggy==1.4.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.20.0
pytest==8.0.1
pytest-django==4.8.0
pytest-docker-compose==3.2.1
pytest-mock==3.12.0
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2024.1
PyYAML==5.4.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
texttable==1.7.0
tomli==2.0.1
tornado==6.4
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
websocket-client==0.59.0
